refactor: reorganize utility functions into dedicated modules for improved maintainability and clarity
