+++ package.json
@@ -12,13 +12,16 @@
   "devDependencies": {
     "@sveltejs/vite-plugin-svelte": "^5.0.3",
     "@tsconfig/svelte": "^5.0.4",
+    "drizzle-kit": "^0.31.1",
     "svelte": "^5.28.1",
     "svelte-check": "^4.1.6",
     "typescript": "~5.8.3",
     "vite": "^6.3.5"
   },
   "dependencies": {
+    "@electric-sql/pglite": "^0.3.2",
     "@tailwindcss/vite": "^4.1.8",
+    "drizzle-orm": "^0.44.2",
     "tailwindcss": "^4.1.8"
   }
 }
