+++ src/lib/currency.ts
@@ -1,0 +1,6 @@
+export function formatCurrency(amount: number, currency: string = 'USD', locale: string = 'en-US'): string {
+  return new Intl.NumberFormat(locale, {
+    style: 'currency',
+    currency: currency,
+  }).format(amount);
+}
