+++ src/lib/date.ts
@@ -1,0 +1,35 @@
+export function formatDate(date: Date): string {
+  return date.toLocaleDateString('en-US', {
+    weekday: 'long',
+    year: 'numeric',
+    month: 'long',
+    day: 'numeric',
+  });
+}
+
+export function isToday(date: Date): boolean {
+  const today = new Date();
+  return date.toDateString() === today.toDateString();
+}
+
+export function getDateRange(centerDate: Date, daysAround: number = 30): Date[] {
+  const dates: Date[] = [];
+  const startDate = new Date(centerDate);
+  startDate.setDate(centerDate.getDate() - daysAround);
+  
+  for (let i = 0; i <= daysAround * 2; i++) {
+    const date = new Date(startDate);
+    date.setDate(startDate.getDate() + i);
+    dates.push(date);
+  }
+  
+  return dates;
+}
+
+export function formatDayOfWeek(date: Date): string {
+  return date.toLocaleDateString('en-US', { weekday: 'short' });
+}
+
+export function isSameDate(date1: Date, date2: Date): boolean {
+  return date1.toDateString() === date2.toDateString();
+}
