--- src/lib/utils.ts
@@ -1,85 +1,0 @@
-export function generateId(): string {
-  return Math.random().toString(36).substr(2, 9);
-}
-
-export function formatCurrency(amount: number, currency: string = 'USD', locale: string = 'en-US'): string {
-  return new Intl.NumberFormat(locale, {
-    style: 'currency',
-    currency: currency,
-  }).format(amount);
-}
-
-export function formatDate(date: Date): string {
-  return date.toLocaleDateString('en-US', {
-    weekday: 'long',
-    year: 'numeric',
-    month: 'long',
-    day: 'numeric',
-  });
-}
-
-export function formatShortDate(date: Date): string {
-  return date.toLocaleDateString('en-US', {
-    month: 'short',
-    day: 'numeric',
-  });
-}
-
-export function isToday(date: Date): boolean {
-  const today = new Date();
-  return date.toDateString() === today.toDateString();
-}
-
-export function isSameMonth(date1: Date, date2: Date): boolean {
-  return date1.getFullYear() === date2.getFullYear() && 
-         date1.getMonth() === date2.getMonth();
-}
-
-export function getMonthDays(date: Date): Date[] {
-  const year = date.getFullYear();
-  const month = date.getMonth();
-  
-  // Get first day of month and last day
-  const firstDay = new Date(year, month, 1);
-  const lastDay = new Date(year, month + 1, 0);
-  
-  // Get first day of week for calendar grid
-  const startDate = new Date(firstDay);
-  startDate.setDate(firstDay.getDate() - firstDay.getDay());
-  
-  // Get last day of week for calendar grid
-  const endDate = new Date(lastDay);
-  endDate.setDate(lastDay.getDate() + (6 - lastDay.getDay()));
-  
-  const days: Date[] = [];
-  const currentDate = new Date(startDate);
-  
-  while (currentDate <= endDate) {
-    days.push(new Date(currentDate));
-    currentDate.setDate(currentDate.getDate() + 1);
-  }
-  
-  return days;
-}
-
-export function getDateRange(centerDate: Date, daysAround: number = 30): Date[] {
-  const dates: Date[] = [];
-  const startDate = new Date(centerDate);
-  startDate.setDate(centerDate.getDate() - daysAround);
-  
-  for (let i = 0; i <= daysAround * 2; i++) {
-    const date = new Date(startDate);
-    date.setDate(startDate.getDate() + i);
-    dates.push(date);
-  }
-  
-  return dates;
-}
-
-export function formatDayOfWeek(date: Date): string {
-  return date.toLocaleDateString('en-US', { weekday: 'short' });
-}
-
-export function isSameDate(date1: Date, date2: Date): boolean {
-  return date1.toDateString() === date2.toDateString();
-}
